requirements.txt
progressbar2
requests
bs4
argparse
re